requirements.txt
flask==3.0.0
python-dotenv==1.0.0
openai==1.12.0
semantic-kernel==1.28.1
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
asgiref==3.7.2
